feat: Implement basic functions

+++ src/main/java/org/example/user_system/controller/AdminController.java
@@ -1,34 +1,48 @@
 package org.example.user_system.controller;
 
+import cn.hutool.core.util.ObjectUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.annotation.Resource;
 import lombok.extern.slf4j.Slf4j;
+import org.example.user_system.context.BaseContext;
 import org.example.user_system.dto.*;
+import org.example.user_system.result.PageResult;
 import org.example.user_system.service.UserService;
 import org.example.user_system.vo.LoginVo;
+import org.example.user_system.vo.UserGetCountVo;
 import org.example.user_system.vo.UserInfoVo;
 import org.example.user_system.vo.UserUpdateInfoVo;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/admin")
-@Tag(name = "通用接口")
+@Tag(name = "管理员接口")
 @Slf4j
 public class AdminController {
 
     @Resource
     private UserService userService;
 
+    /**
+     * 管理员登录
+     * @param loginDto 用户登录信息
+     * @return LoginVo
+     */
     @PostMapping("/login")
     @Operation(summary = "管理员登录")
-    public LoginVo login(@RequestBody LoginDto loginDto) {
+    public LoginVo login(@RequestBody @Validated LoginDto loginDto) {
         log.info("管理员登录 loginDto = {}", loginDto);
 
         return userService.adminLogin(loginDto);
     }
+
 
     @PostMapping("/add")
     @Operation(summary = "新增用户")
@@ -41,6 +55,10 @@ public class AdminController {
     @DeleteMapping("/delete")
     @Operation(summary = "删除用户")
     public void delete(@RequestParam List<Integer> ids) {
+        if(ObjectUtil.isEmpty(ids)) {
+            throw new RuntimeException("请选择需要删除的用户");
+        }
+
         log.info("删除用户 ids = {}", ids);
 
         userService.delete(ids);
@@ -48,7 +66,11 @@ public class AdminController {
 
     @PutMapping("/update/password")
     @Operation(summary = "修改密码")
-    public void updatePassword(UserUpdatePasswordDto userUpdatePasswordDto) {
+    public void updatePassword(@RequestBody UserUpdatePasswordDto userUpdatePasswordDto) {
+        if(ObjectUtil.isEmpty(userUpdatePasswordDto.getUserId())) {
+            userUpdatePasswordDto.setUserId(BaseContext.getCurrentUser().getUserId());
+        }
+
         log.info("修改密码 userUpdatePasswordDto = {}", userUpdatePasswordDto);
 
         userService.updatePassword(userUpdatePasswordDto);
@@ -56,7 +78,11 @@ public class AdminController {
 
     @PutMapping("/update/role")
     @Operation(summary = "更新用户角色")
-    public void updateRole(UserUpdateRoleDto userUpdateRoleDto) {
+    public void updateRole(@RequestBody UserUpdateRoleDto userUpdateRoleDto) {
+        if(ObjectUtil.isEmpty(userUpdateRoleDto.getUserId())) {
+            throw new RuntimeException("请选择你需要修改的用户");
+        }
+
         log.info("更新用户角色 userUpdateRoleDto = {}", userUpdateRoleDto);
 
         userService.updateRole(userUpdateRoleDto);
@@ -64,7 +90,11 @@ public class AdminController {
 
     @PutMapping("/update/info")
     @Operation(summary = "修改用户个人信息")
-    public UserUpdateInfoVo updateInfo(UserUpdateInfoDto userUpdateInfoDto) {
+    public UserUpdateInfoVo updateInfo(@RequestBody UserUpdateInfoDto userUpdateInfoDto) {
+        if(ObjectUtil.isEmpty(userUpdateInfoDto.getId())) {
+            userUpdateInfoDto.setId(BaseContext.getCurrentUser().getUserId());
+        }
+
         log.info("修改用户个人信息 userUpdateInfoDto = {}", userUpdateInfoDto);
 
         return userService.updateInfo(userUpdateInfoDto);
@@ -73,6 +103,10 @@ public class AdminController {
     @GetMapping("/get/info")
     @Operation(summary = "获取用户详细信息")
     public UserInfoVo getInfo(Integer id) {
+        if(ObjectUtil.isEmpty(id)) {
+            id = BaseContext.getCurrentUser().getUserId();
+        }
+
         log.info("获取用户详细信息 id = {}", id);
 
         return userService.getInfo(id);
@@ -80,17 +114,30 @@ public class AdminController {
 
     @GetMapping("/get/count")
     @Operation(summary =  "获取用户总数")
-    public Long getCount() {
+    public UserGetCountVo getCount() {
         log.info("获取用户总数");
 
-        return userService.count();
+        return UserGetCountVo.builder().num(userService.count()).build();
     }
 
     @GetMapping("/query")
     @Operation(summary = "根据条件查询用户列表")
-    public List<UserInfoVo> query(UserQueryDto userQueryDto) {
+    public PageResult<UserInfoVo> query(UserQueryDto userQueryDto) {
+        if (ObjectUtil.isEmpty(userQueryDto.getPage())) {
+            userQueryDto.setPage(1);
+        }
+        if (ObjectUtil.isEmpty(userQueryDto.getSize())) {
+            userQueryDto.setSize(10);
+        }
+
         log.info("根据条件查询用户列表 userQueryDto = {}", userQueryDto);
 
         return userService.query(userQueryDto);
+    }
+
+    @DeleteMapping("/logout")
+    @Operation(summary = "登出")
+    public void logout() {
+        userService.logout();
     }
 }
